Issue222.abandon dot net framework

+++ SimpleBrowser.UnitTests/OfflineTests/FileUri.cs
@@ -18,9 +18,23 @@ namespace SimpleBrowser.UnitTests.OfflineTests
         [Test]
         public void CanLoadHtmlFromFile()
         {
-            FileInfo f = new FileInfo(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, @"SampleDocs\movies1.htm"));
-            string uri = string.Format("file:///{0}", f.FullName);
-            uri = uri.Replace("\\", "/");
+            FileInfo f = null;
+            string uri = string.Empty;
+            if (Environment.OSVersion.Platform == PlatformID.Win32NT)
+            {
+                f = new FileInfo(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, @"SampleDocs\movies1.htm"));
+                uri = string.Format("file:///{0}", f.FullName);
+                uri = uri.Replace("\\", "/");
+            }
+            else if (Environment.OSVersion.Platform == PlatformID.Unix)
+            {
+                f = new FileInfo(Path.Combine("home", AppDomain.CurrentDomain.BaseDirectory, @"SampleDocs/movies1.htm"));
+                uri = string.Format("file://{0}", f.FullName);
+            }
+            else
+            {
+                throw new NotImplementedException("Please write unit tests for this unknown platform. (MacOS?)");
+            }
 
             Browser b = new Browser();
             b.Navigate(uri);
@@ -30,7 +44,8 @@ namespace SimpleBrowser.UnitTests.OfflineTests
         [Test]
         public void CanLoadHtmlFromFilesWithAbsolutePath()
         {
-            if (Directory.Exists("C:\\Windows\\Temp"))
+            if (Environment.OSVersion.Platform == PlatformID.Win32NT &&
+                Directory.Exists("C:\\Windows\\Temp"))
             {
                 File.Copy(
                     Path.Combine(AppDomain.CurrentDomain.BaseDirectory, @"SampleDocs\movies1.htm"),
@@ -48,6 +63,21 @@ namespace SimpleBrowser.UnitTests.OfflineTests
 
                 b.Navigate("file://\\c|\\Windows\\Temp\\movies1.htm");
                 Assert.AreEqual(b.Select("ul#menu>li").Count(), 3);
+
+                File.Delete(@"C:\Windows\Temp\movies1.htm");
+            }
+            else if (Environment.OSVersion.Platform == PlatformID.Unix &&
+                Directory.Exists("/tmp"))
+            {
+                File.Copy(
+                    Path.Combine("home", AppDomain.CurrentDomain.BaseDirectory, @"SampleDocs/movies1.htm"),
+                    @"/tmp/movies1.htm", true);
+
+                Browser b = new Browser();
+                b.Navigate("file:///tmp/movies1.htm");
+                Assert.AreEqual(b.Select("ul#menu>li").Count(), 3);
+
+                File.Delete(@"/tmp/movies1.htm");
             }
         }
     }
